RFID Access control system
Powered by passive poe, this is an early prototype. Eventually it will
send the read tag back to the server over mqtt and receive a response
depending on the confirmation status. I.E. Response was yes, turn led
green and open relay. Or the response was no so turn the led red and
keep the door shut. Also planning to add support for motion sensor and
contact sensor and maybe a release button for egress. Hoping to tie in
lock support with HomeKit through homebridge-mqtt and homebridge.